Enable offline gradle builds with --offline flag
Before deploying to the nano, a local repo needs to be built with
./gradlew sync. This was elided from being a dependency, as it adds a
couple seconds to rapid redeployments. Just remember to resync when new
dependencies are added.

Ivypot recursively downloads all dependencies and supplies them as a
local repo. A lot of methods and tools were explored, but they were
either convoluted (copying .cache?) or ineffective (flatDirs don't fill
out recursively). This does not resolve the initial gradlew download --
that still needs to get a first run when connected to the internet.

Version 8, as of the most recent build (v8.1.1), breaks this. DO NOT
UPDATE TO GRADLE VERSION 8.